refactor(tests): match test names with packages under libs
