--- src/test_list.py
@@ -1,6 +1,0 @@
-test_list: list[int] = None
-
-if __name__ == "__main__":
-    test_list = []
-    test_list.append(10)
-    print(test_list)
